[BUGFIX] Check if HBNB_ENV exists in os.environ before trying to access in db_storage

+++ models/engine/db_storage.py
@@ -40,7 +40,7 @@ class DBStorage:
         self.reload()
 
         # drop all tables if HBNB_ENV == 'test'
-        if os.environ['HBNB_ENV'] == 'test':
+        if 'test' in os.environ and os.environ['HBNB_ENV'] == 'test':
             Base.metadata.drop_all(self.__engine)
 
     def all(self, cls=None):
